refactor(opengl_shader): moved static functions from lv_gltf_view_shader into lv_opengl_shader_manager.c
